+++ src/main/java/com/senity/waved/domain/paymentRecord/entity/PaymentRecord.java
@@ -1,8 +1,10 @@
 package com.senity.waved.domain.paymentRecord.entity;
 
 import com.senity.waved.common.BaseEntity;
-import com.senity.waved.domain.paymentRecord.dto.response.PaymentRecordResponseDto;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
@@ -36,15 +38,6 @@ public class PaymentRecord extends BaseEntity {
                 .memberId(memberId)
                 .myChallengeId(myChallengeId)
                 .groupTitle(groupTitle)
-                .build();
-    }
-
-    public static PaymentRecordResponseDto getPaymentResponse(PaymentRecord paymentRecord) {
-        return PaymentRecordResponseDto.builder()
-                .groupTitle(paymentRecord.getGroupTitle())
-                .status(paymentRecord.getPaymentStatus())
-                .deposit(paymentRecord.getDeposit())
-                .createDate(paymentRecord.getCreateDate())
                 .build();
     }
 }
