+++ src/main/java/com/senity/waved/domain/paymentRecord/repository/PaymentRecordRepository.java
@@ -1,10 +1,10 @@
 package com.senity.waved.domain.paymentRecord.repository;
 
 import com.senity.waved.domain.paymentRecord.entity.PaymentRecord;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface PaymentRecordRepository extends JpaRepository<PaymentRecord, Long> {
-    List<PaymentRecord> getPaymentRecordByMemberId(Long memberId);
+    Page<PaymentRecord> getPaymentRecordByMemberId(Long memberId, Pageable pageable);
 }
