+++ src/main/java/com/senity/waved/domain/review/repository/ReviewRepository.java
@@ -1,11 +1,11 @@
 package com.senity.waved.domain.review.repository;
 
 import com.senity.waved.domain.review.entity.Review;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> getReviewByMemberId(Long memberId);
-    List<Review> findByChallengeId(Long challengeId);
+    Page<Review> getReviewByMemberId(Long memberId, Pageable pageable);
+    Page<Review> findByChallengeId(Long challengeId, Pageable pageable);
 }
